+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.3.0-Beta
+
+- Feat: add Windows-only scheduled notification, developer settings page (playground), general fixes
+- Feat: added Scheduled Notifications settings page /Front-end only: missing notification logic
+- Refractor: MaterialCardList - Added MaterialCard single and list methods, deprecated materialCardList()
+- Chore: delete empty file
+- Refractor: MaterialCardList - materialCardList() to MaterialCard.list()
+- Fix: MaterialCardList - card not appearing on surfaceContainerHigh
+- Fix: MaterialCardList - Material ripple not clipping with the card
+- Fix: MaterialCard/MaterialCardList - Switched background to surfaceContainerHigh
+- Fix: MaterialConnectedButtonGroup - button cancelling whole selection when deselecting only one item.
+
 ## 0.2.0-Beta
 
 - Feat: added Permissions settings page